Changelog 1. 적 체력 2. 적 움직임 3. 적 공격 4. 콤보 공격 5. 카메라 제어

+++ Assets/Scripts/ILivingEntity.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface ILivingEntity
 {
-    void TakeDamage(float damage, GameObject instigator);
+    void TakeDamage(int damage, GameObject instigator);
 }
